auth-service/config.py
import os
from dotenv import load_dotenv

load_dotenv(override=False)

class Config:
    SECRET_KEY = os.getenv("SECRET_KEY")
    SQLALCHEMY_DATABASE_URI = os.getenv("DATABASE_URL")
    SQLALCHEMY_TRACK_MODIFICATIONS = False

    SQLALCHEMY_ENGINE_OPTIONS = {
        "pool_pre_ping": True,
        "pool_recycle": 280,
        "connect_args": {"connect_timeout": 5}
    }
    
    JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY")