auth-service/requirements.txt
flask 
flask_sqlalchemy 
psycopg2-binary 
flask-bcrypt 
flask-jwt-extended 
python-dotenv
flask-migrate
gunicorn